requirements.txt
playwright==1.50.0
requests==2.32.3
PyYAML==6.0.2
